+++ ikii-back-microservices/ikii-back-products/src/main/java/mx/ikii/products/repository/productbusiness/IProductBusinessRepositoryCustom.java
@@ -1,0 +1,11 @@
+package mx.ikii.products.repository.productbusiness;
+
+import java.util.List;
+
+import mx.ikii.commons.persistence.collection.ProductBusiness;
+
+public interface IProductBusinessRepositoryCustom {
+	
+	List<ProductBusiness> searchByKeywords(String keyword);
+
+}